fixed chat bug that was truncating some messages, due to using a uint16_t to store a local size. Totally backward compatible. This fix might also affect blog messages.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/branches/v0.5.0@2879 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/serialiser/rstlvbase.h
@@ -209,7 +209,7 @@ const uint16_t TLV_TYPE_IMAGE         = 0x1060;
 
 
 /**** Basic TLV Functions ****/
-uint16_t GetTlvSize(void *data);
+uint32_t GetTlvSize(void *data);
 uint16_t GetTlvType(void *data);
 bool     SetTlvBase(void *data, uint32_t size, uint32_t *offset, uint16_t type, uint32_t len);
 bool     SetTlvSize(void *data, uint32_t size, uint32_t len);
